card entity/repo ready and tested

+++ src/main/java/org/if22b179/apps/mtcg/entity/Card.java
@@ -6,7 +6,7 @@ import lombok.Data;
 public class Card {
     private String id; // Eindeutige Identifikationsnummer der Karte
     private String name;
-    private int damage;
+    private double damage;
     private ElementType elementType;
     private CardType cardType;
     private String ownerUsername; // Benutzername des Besitzers der Karte
